# Conflicts:
#	backend/.gradle/8.2.1/executionHistory/executionHistory.bin
#	backend/.gradle/8.2.1/executionHistory/executionHistory.lock
#	backend/.gradle/8.2.1/fileHashes/fileHashes.bin
#	backend/.gradle/8.2.1/fileHashes/fileHashes.lock
#	backend/.gradle/8.2.1/fileHashes/resourceHashesCache.bin
#	backend/.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	backend/.gradle/file-system.probe
#	backend/build/classes/java/main/com/bitc/finalproject/entity/BoardEntity.class
#	backend/build/tmp/compileJava/previous-compilation-data.bin
#	frontend/node_modules/.cache/.eslintcache
#	frontend/node_modules/.cache/default-development/0.pack
#	frontend/node_modules/.cache/default-development/1.pack
#	frontend/node_modules/.cache/default-development/12.pack
#	frontend/node_modules/.cache/default-development/13.pack
#	frontend/node_modules/.cache/default-development/14.pack
#	frontend/node_modules/.cache/default-development/15.pack
#	frontend/node_modules/.cache/default-development/2.pack
#	frontend/node_modules/.cache/default-development/4.pack
#	frontend/node_modules/.cache/default-development/5.pack
#	frontend/node_modules/.cache/default-development/7.pack
#	frontend/node_modules/.cache/default-development/8.pack
#	frontend/node_modules/.cache/default-development/9.pack
#	frontend/node_modules/.cache/default-development/index.pack
#	frontend/node_modules/.cache/default-development/index.pack.old
#	frontend/src/index.js

+++ backend/src/main/java/com/bitc/finalproject/controller/BoardControl.java
@@ -36,14 +36,15 @@ public class BoardControl {
     public Object boardInsert(BoardEntity boardEntity) throws Exception {
         boardService.writeBoard(boardEntity);
 
-        System.out.println(boardEntity);
-
         return "redirect:/board";
     }
 
     @RequestMapping(value = "/board/{boardPk}", method = RequestMethod.DELETE)
-    public Object boardDelete(@PathVariable("boardPk") int boardPk) throws Exception {
-        boardService.deleteBoard(boardPk);
+    public Object boardDelete(@PathVariable("boardPk") int boardPk, @RequestParam("boardWriterId") String boardWriterId) throws Exception {
+
+        if (boardWriterId.equals("testUserId")) {
+            boardService.deleteBoard(boardPk);
+        }
 
         return "redirect:/board";
     }
